Additional RINEX .nav files for parser unit tests
Additional RINEX .nav files in versions 2.00 through 3.04, inclusive.

Most of these were taken directly from the examples in the indicated versions of the RINEX spec itself. The one from ANKR00TUR (RINEX 3.03) was corrupted on the IGS server, so I thought it would make an interesting (negative/adversarial) test case. Also including a cleaned-up version of that same file, as it's the only one I have in 3.03 format.